+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### `jsonschema-generator`
+#### Added
+- New `Option.FLATTENED_SUPPLIERS` to unwrap the supplied type. `Supplier<T>` would thus be a `type: "T"`.
 
 ## [4.24.3] - 2022-05-03
 ### `jsonschema-generator`